Fix clippy warnings

+++ src/sources/util/framestream.rs
@@ -66,7 +66,7 @@ const FSTRM_CONTROL_FIELD_CONTENT_TYPE_LENGTH_MAX: usize = 256;
 /// it should release its permit (and try to obtain a new one) allowing other connections to read.
 /// It is very short because any incoming data will avoid this timeout,
 /// so it mainly prevents holding permits without consuming any data
-const PERMIT_HOLD_TIMEOUT_MS: usize = 10;
+const PERMIT_HOLD_TIMEOUT_MS: u64 = 10;
 
 pub type FrameStreamSink = Box<dyn Sink<Bytes, Error = std::io::Error> + Send + Unpin>;
